Initial commit
Primary functionality is implemented, launching the executable will set a pre-configured plan named 'Insomnia' as active, which should be configured to prevent the computer from going to sleep ever.

Clicking the button the main window will turn off the monitor without putting the device to sleep and closing the window will restore the initial state.